Change settings format to use Qt Arrays
Before this, settings for the list of tags would be stored in a .conf
file as QVariants that were wrappers around Python objects. These are
unreadable, and led to issues whenever the API changed. By storing them
in a format that QSettings expects, they become readable in the file,
and less likely to break other parts of the code.