[prefs] Put tags into one list
Rather than needing to manage three lists, and moving between them,
combine into one.
To know which tags should be displayed on the table, add a field to the
TagInfo object called `show_in_table`, set to False by default.
Remove the properties from prefs for custom_tags and standard_tags.
Keep the property for file tags, as it allows quick population of the
table, but rename it to table_tags.