[prefs] Store table_tags in a cached list
A call to table_tags was leading to a call to info_tags which calls
getArray. If this happens or every single cell in the table, it's a
useless and significant slow down. Instead, store the info that would be
retrieved once, and invalidate it when it needs to be changed.