refactor[prefs]: Create new split prefs dialog
The current way of updating settings is unintuitive: update some in the
prefs dialog, some in the tag dialog, and then they're only committed
when Ok is clicked inside the prefs dialog.

Instead, make the prefs dialog an entry way to different dialogs for the
different setting types, and update the settings for those types when
the child dialog is accepted.