fix[statusbar]: Stop message from deleting when clicking menus
The statusbar message that displays the number of songs is overridden by
the empty statusTip messages that get called from the menuAction()
function when selecting a menu item.

There are two ways to fix this: add an eventFilter to the statusbar that
filters out all QStatusTipEvents, or add QLabels explicitly, and don't
use the default message function.

The second one works better because it also allows us to move the
version message from the right to the left, which was the original
desired behaviour.