refactor[edit-dialog]: Create abstract group widget
It was confusing that the LineEdit needed the functions addRow,
removeRow and moveRow, because it inherited from an abstract class that
required those.
This was done because the createButtons method needed those functions to
exist to map the signals from the buttons to those functions as slots.

This creates a new class that the table widget and list widget can
inherit from, that defines those methods.
It also removes the connections from the createButtons method inside the
original abstract class, and overrides it inside the group abstract
class.