refactor[utils]: Move utils into related folders
Done to prevent import cycles when needing a util function in a class
that another util group uses.
In particular, view_utils is dependent on SongRepository, which is
dependent on Song.
But Song was now dependent on tag_utils.
So when initialising the utils module, the import cycle would happen.