fix[edit-dialog]: Fix table not resetting properly if row deleted
The number of rows was needs to be reset when the data in a table is
displayed, otherwise it will only populate the existing number of rows
with the original data, and any original data that is in a higher row
number won't be added to the table. This also has the opposite effect,
where there would be empty rows if the current data has more rows than
the original reset data.