refactor[edit-dialog]: Explicitly mark as abstract and use generics
There are two warnings this resolves:

1. `reportImplicitAbstractClass`: This happens because the abstract
classes weren't inheriting from ABC, or declaring metaclass=ABCMeta.
This was because they were declaring a joint metaclass ABCQMeta. But
inheriting from ABC here doesn't lead to any issues.

2. `reportIncompatibleMethodOverride`: This happened because the
subclasses of the abstract widget deal with specific types, but the
abstract class declares a general type. Specifically, this happens with
the setter, and it's because of contravariance. Because the concrete
widgets always deal with a specific type, it's more convenient to just
use generics that can be declared for the class.