refactor[edit-dialog]: Change to explicit inner functions
Rather than lambdas and confusing inner functions in the updateTag
function, create a new class that has the logic for how to perform the
change, given a specific function and sequence of items.

Define the function names explicitly, so that it's more clear what their
intentions are, and easier to create instances of that mapping class.