tests: Ensure more coverage
Ensures that most of the code is actually run by tests, for the existing
tests. Code was removed for cases that couldn't actually happen, and
documentation expanded for edge cases.

Also moves to using a temporary directory for mp3 files. This means only
one file needs to actually be stored, as we can make copies of that.