refactor(edit): Move properties, init, and clear to abstract
Rather than having the same logic in each of the concrete classes, this
logic can be made a final method in the abstract class.

The only complication is that `_initValue` and `_clearValue` need to
know the type for the empty value. This was solved by creating a new
property called `empty`.