refactor(ui): Remove ui files
The Python code is automatically generated from the `.ui` files,
but this means it doesn't have type-hints, or any other information.

For better linting and type-checking, the code has been rewritten and
improved to not use any ui files.

It also introduces a mixin (`RowOperationMixin`) for row operations.
This is done because the logic for moving, adding, and removing rows
from the widgets in both prefs dialogs and edit dialogs is very similar.
This mixin creates the signals, sets the icon, and defines instance
attributes for the buttons that perform the operations.
It is an abstract class that requires that the `_addRow`, `_removeRow`
and `_moveRows` functions are defined.

* refactor(about): Remove ui files
* refactor(dialog-prefs): Remove ui files
* refactor(dialog-prefs-tag-add): Remove ui files
* refactor: Create RowOperationMixin
* refactor(dialog-prefs): Remove ui files
* refactor(main-window): Remove ui files