fix: EditWidgetFactory convert string to list
It's possible (theoretically) that the data for an EditWidget is for a
tag that displays multiple data, but it's being passed a single string.
A string doesn't have a `copy` method, so that wouldn't work, and also
limits adding multiple fields. The solution is to convert any strings
for multiple data tags into list of strings, before creating the widget.