tests: Fix relative path for assets
Depending on how pytest is called, it could fail to find the
`test_song.mp3` file in the assets folder. The conftest config was
assuming pytest would always be run from the project root, but this
isn't the case for PyCharm. So, use the path of the actual conftest file
to find the song, instead.