refactor: Move meta file
The ABCQMeta metaclass is needed whenever there are abstract classes
that are also QObjects. This isn't always going to just be the case with
GUI classes, if we start using QObjects to get signals in other classes.
So move this metaclass file into a "core" folder that is separate from
GUI.