refactor(TagValue): Make AbstractTagValue a QObject
This allows it to emit signals when the join character updates.

The idea is that it should receive a signal that tells it when it has
changed, from the settings object. Then it emits a signal that can be
caught if it truly has changed, which could then be used by the model to
know that the data it's displaying should be updated.