refactor(song-repo): Simplify interface and rename
It seems silly to have `addFile`, `addSong`, `addFiles`, and `getSong`
methods. The first three have been combined into a single `add` method,
and `getSong` has been removed, in favour of just using `__getitem__`.

It changes `__getitem__` to raise an IndexError if the index is invalid,
to be more in line with how that normally works.