fix(song-repo): Allow adding multiple songs with no file
Problem:
Adding a song that is not related to a specific file should be able to
be done multiple times. This is useful for testing. At the moment,
because it does a check for whether the file attribute for the song is
already present - which it will be, with a value of None - before it
adds a song, only one empty song can be added.

Solution:
Do an explicit check for whether song.file is not None in the
`__contains__` method.