refactor(song)!: Use TagValue for song and store internally
Instead of doing the storing and updating outside, and then calling the
song with the updates, store the updates in the song object itself.

This affects all objects that interact with Song, particularly related
to getting and setting tags. It is a breaking change, because it no
longer uses SongTag, and no longer accepts bare values - they need to be
converted to AbstractTagValue instances using TagValueFactory.